feat: widen on-the-move fix interval to uint16 (1–65535 seconds)
uint8_t capped acf= at 255 seconds (~4 min), which is unnecessarily
restrictive now that the unit is seconds. Change _onTheMoveFixInterval
to uint16_t so the full range 1–65535 s is available, consistent with
_defaultFixInterval and _gpsFixTimeout which are already uint16_t.

- Config.h: field uint8_t -> uint16_t, getter return type updated
- Config.cpp: command handler set/show_uint8 -> set/show_uint16
- Config.cpp: bump EEPROM header 0xBEF1 -> 0xBEF2 (struct size changed)
- Readme.md: document the 1–65535 range for acf=

+++ SodaqOneTracker_v2/Config.h
@@ -51,7 +51,7 @@ struct ConfigParams
 
     uint8_t _accelerationPercentage;
     uint8_t _accelerationDuration;
-    uint8_t _onTheMoveFixInterval;
+    uint16_t _onTheMoveFixInterval; // seconds (1–65535)
     uint8_t _onTheMoveTimeout;
 
     int8_t _temperatureSensorOffset;
@@ -100,7 +100,7 @@ public:
 
     uint8_t getAccelerationPercentage() const { return _accelerationPercentage; }
     uint8_t getAccelerationDuration() const { return _accelerationDuration; }
-    uint8_t getOnTheMoveFixInterval() const { return _onTheMoveFixInterval; }
+    uint16_t getOnTheMoveFixInterval() const { return _onTheMoveFixInterval; }
     uint8_t getOnTheMoveTimeout() const { return _onTheMoveTimeout; }
 
     int8_t getTemperatureSensorOffset() const { return _temperatureSensorOffset; };
